Refactor WorkloadVersion controller by making EvaluationHandler injectable Signed-off-by: geoffrey1330 <israelgeoffrey13@gmail.com>
